Added workflow variable to master_xml.py
Using workflow as a parameter when run the master.xml stylesheets to indicate that something is a website.  This is done so that a format identification of warc can manually be added to the master.xml, since FITS only identifies them as gzip files.

If it isn't a website, there is no 4th argument when this script is run and so it assigns a default value of general, which has no affect.