Work with BMAC identifiers
BMAC uses this general-aip workflow for web aips. Needed to update make_preservationxml() to leave department unchanged for bmac and update the regex in the xsl files for bmac collection and file identifiers.

Did as a proof of concept with approximate bmac aip id structure for web aips. Will need to update the regex once the  aip id structure is finalized.