Merge the two fits to preservation stylesheets
Before, had one stylesheet for single file and one for multiple files. Used Russell AV AIP script as a model to combine these into a single stylesheet where it counts the number of files and has a few variations (representation or file as the object category in AIP and if it includes the filelist) based on that count.