Add aip_folder to metadata.csv
Using aip_folder in metadata.csv to match the folder to its metadata without requiring it to be renamed to any of the metadata values. Previously, was requiring it to be the title. This gives us more flexibility.

Also made the error message for if the metadata.csv is not found more specific.