metadata csv count matches folder count
The same AIPs need to be in the metadata csv and AIPs directory folder. As a quick check, verify that the counts are the same. If time, we'll add verifying the folder titles match what is in the csv.

Also renamed variables open_aip and read_aip to open_metadata and read_metadata to be more descriptive.