Add aip-id to fits error file name
To be consistent with other metadata files, start the log of fits tool errors with the AIP ID.

Also moved creating the log to after all of the first script checks, since they aren't included in the log. That way, don't get a log with just the start date and no other information if any check quits the script.