Add check_metadata_csv()
Moves the existing metadata check for header order to this function and adds a comparison of the folder names in the csv and the AIPs directory. This replaces comparing a count of the two since it would be possible to get errors in both the csv and AIPs directory which cancelled each other out.