Rename script log
No longer need the date in the filename. Want to be able to run the script on multiple days as fix errors and add them to the log. Timestamps within the log give necessary time information.

Changed from script_log to general_aip_script_log at the same time to be more descriptive.