Attempt to add to existing log without header
Moved addition of header to the log() function and trying to only add a header row if the log file doesn't already exist. Not working correctly though - able to add new log information to a log left from a previous run of the script as desired but will add a new header row every time.