Change how log is made
Using a dictionary to store log information, instead of a list, so that it is more clear what information is being saved within the functions. log() converts the dictionary values to a list for saving to the log. Also updated the log messages to be more descriptive in case the data ever ends up in the wrong row.