Save preservation.xml validation output
Makes a file named with the aip-id inside the error folder for preservation.xml validation and saves the output from the validation (each line on its own row, double spaced) for review.

Put in the script log initially but lose information due to cell limits and harder to read anyway.