test_make_output_directories.py: better comparison
Using the change in directories within the parent folder instead of testing for if the three exist.

Also added tearDown to remove the output directories rather than including them within the test so they still work even if the assert has an error, to prevent a problem with the next test.