Actually fix path errors in tests
Refactoring when moving the script out of the "scripts" folder to the main repo changed the paths to the output folders, errors folder, and aip_log in aip_functions.py. Originally, I thought the problem was the tests referring to a parent folder in the path. It was actually that the functions had "" instead of ".." in the path.